Fix floating-point precision

+++ src/main/java/pl/patkis/zielonykod/transactions/Account.java
@@ -1,8 +1,10 @@
 package pl.patkis.zielonykod.transactions;
 
+import java.math.BigDecimal;
+
 public class Account {
     public String account;
     public int debitCount;
     public int creditCount;
-    public double balance;
+    public BigDecimal balance;
 }
