+++ src/main/java/pl/patkis/zielonykod/transactions/Transaction.java
@@ -1,7 +1,9 @@
 package pl.patkis.zielonykod.transactions;
 
+import java.math.BigDecimal;
+
 public class Transaction {
     public String debitAccount;
     public String creditAccount;
-    public double amount;
+    public BigDecimal amount;
 }
